Move section outside of candidate interface
we have a spec/lib/namespace_spec that checks namespaces intrusions
