+++ app/forms/support_interface/editable_until_form.rb
@@ -8,11 +8,11 @@ module SupportInterface
     validates_with ZendeskUrlValidator
 
     def non_editable_sections
-      CandidateInterface::Section.non_editable.unshift(science_gcse).flatten.compact
+      Section.non_editable.unshift(science_gcse).flatten.compact
     end
 
     def science_gcse
-      CandidateInterface::Section.all.find { |section| section.id == :science_gcse }
+      Section.all.find { |section| section.id == :science_gcse }
     end
 
     def sections
